+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HumHub Changelog
 ---------------------------------
 - Enh #7307: Improve request scheme detection
 - Fix #7308: Fix Hungarian symbol error in `Open Sans` font 
+- Fix #7312: Fix obsolete nonce value on Live Push driver
 
 1.17.0-beta.2 (November 12, 2024)
 ---------------------------------